server: move controller startup into controllers.go
